Finished algorithm
Algorithm probably needs bigger cells so AI doesn't stumble on obsticles. Any kind of obsticle can be put on track and it needs to be tagged as "Obsticle". After that all the cells need to be deleted(manually, for now) and then pathfinder needs to be enabled and algorithm will detect obsticles and where track is is now walkable. There I can implement so AI can recognize other parts of track, like dirt or grass or something else. Start and finish aren't still solved